jobspy_llm_letters/requirements.txt
# JobSpy LLM Letters - Requirements
# Core job scraping library
python-jobspy>=1.1.82

# Data processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Configuration and file handling
pyyaml>=6.0,<7.0
python-dotenv>=1.0.1,<2.0.0

# Progress bars and UI
tqdm>=4.66.0,<5.0.0

# LLM API clients
openai>=1.40.0,<2.0.0

# Optional: Alternative LLM providers (uncomment as needed)
# anthropic>=0.25.0  # For Claude API
# google-generativeai>=0.5.0  # For Gemini API

# Optional: Enhanced logging and monitoring
# loguru>=0.7.0  # Better logging
# rich>=13.0.0   # Rich console output

# Development and testing (install with: pip install -r requirements-dev.txt)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0

# System requirements:
# - Python 3.10 or higher
# - OpenAI API key (set as OPENAI_API_KEY environment variable)
# - Stable internet connection for job scraping